[JAVA-47201] Created new module json-conversion-2 and moved code from json-conversion
