--- json-modules/README.md
@@ -1,3 +1,0 @@
-## JSON
-
-This module contains modules about JSON.
