+++ json-modules/json-conversion-2/src/main/java/com/baeldung/includenullinjson/Customer.java
@@ -1,0 +1,27 @@
+package com.baeldung.includenullinjson;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class Customer {
+    @JsonProperty
+    private final String name;
+    @JsonProperty
+    private final String address;
+    @JsonProperty
+    private final int age;
+
+    public Customer(String name, String address, int age) {
+        this.name = name;
+        this.address = address;
+        this.age = age;
+    }
+
+    @Override
+    public String toString() {
+        return "Customer{" +
+                "name='" + name + '\'' +
+                ", address='" + address + '\'' +
+                ", age=" + age +
+                '}';
+    }
+}
