+++ json-modules/json-conversion-2/src/main/java/com/baeldung/jsonnodetocollection/CustomPersonListDeserializer.java
@@ -1,13 +1,13 @@
 package com.baeldung.jsonnodetocollection;
 
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
 import com.baeldung.jsonnodetocollection.dto.Person;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 
 public class CustomPersonListDeserializer extends JsonDeserializer<List<Person>> {
     @Override
