+++ json-modules/json-conversion-2/src/main/java/com/baeldung/jsonnodetocollection/JsonNodeConversionUtil.java
@@ -1,15 +1,15 @@
 package com.baeldung.jsonnodetocollection;
 
-import com.baeldung.jsonnodetocollection.dto.Person;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import com.baeldung.jsonnodetocollection.dto.Person;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
  public class JsonNodeConversionUtil {
