+++ json-modules/json-conversion-2/src/main/java/com/baeldung/jsonnodetocollection/dto/Person.java
