+++ json-modules/json-conversion/src/main/java/com/baeldung/jsonarraytolist/ConvertJsonArrayToList.java
@@ -9,8 +9,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
-import entities.Product;
-
 public class ConvertJsonArrayToList {
 
     public List<Product> convertJsonArrayUsingGsonLibrary(String jsonArray) {
