+++ json-modules/json-conversion/src/main/java/com/baeldung/jsonarraytolist/Product.java
@@ -1,4 +1,4 @@
-package entities;
+package com.baeldung.jsonarraytolist;
 
 public class Product {
